A lot of progress
Made it so you can drop items from the inventory and they would get constructed as gameobjects and dropped into the world.

I created an effects class for the Item Equipment class. Each piece of equipment now stores a list of all effects you want it to use when the equipment is clicked on in the inventory. The list of effects will each call a castEffect function that does something. You can add or remove effects as you choose, it's completely modular.

I added some better icons,.
I added a material to the base item class, along with a mesh. These are used for when it's dropped and created as an item in the world.

Still need to do:
Add a custom editor window for creating items/equipment.
Wrap it all as a plugin.